codeshin-backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
numpy==2.2.4
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
